Properly handle plain Error objects (#28)
# Description

The `error` logger method expects all errors to have a `status` property. But there are cases when it can be invoked with plain Error objects, for instance by plugins in `try..catch` statements.

Since plain Error objects do not have a `status` property, they are stringified. Problem is: the stringified version of an Error object is always `{}`

This PR also add many missing unit tests about the error logger method.

# Other changes

* Update unit test dependencies (winston itself will have to wait as there are many breaking changes from v2 to v3)
* Use the `dot` mocha reporter, more concise than the default one
* Fix the version of `winston-syslog` because the new 2.1.0 minor is not compatible with winston 2